RWA-861 [SH] Adding unit tests to increase pitest coverage
